Revert "Merge branch 'refactor/reactQuery' into dev"
This reverts commit 5be94fd66a5e0d2f8f6f9403f97c3ff72fa92a38, reversing
changes made to e07ccab6ff7cc57a8f75313df55b57611be1940f.
