+++ src/pages/AlarmListPage.tsx
@@ -7,7 +7,7 @@ import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { useLocation } from 'react-router-dom';
 import { ItemProps } from 'stories/listbox/type';
-import { getListApi } from '../service/AlarmService';
+import { getListApi } from '../apis/alarm';
 
 const StyledDatePicker = styled(DatePicker)(({ theme }: { theme: any }) => ({
     '& .MuiInputBase-root': {
@@ -35,7 +35,7 @@ function getSevenDaysAgoDate() {
     return `${year}-${month}-${day}`;
 }
 
-const AlarmListPage = () => {
+export const AlarmListPage = () => {
     const { from, to } = useLocation().state;
     const [stDate, setStDate] = useState<string>(getSevenDaysAgoDate());
     const [enDate, setEnDate] = useState<string>(getCurrentDate());
@@ -89,5 +89,3 @@ const AlarmListPage = () => {
         </Box>
     );
 };
-
-export default AlarmListPage;
